make report form
