+++ src/routes/Home.tsx
@@ -1,7 +1,8 @@
 import React, { useEffect, useState } from 'react'
 import wfhdog from '../assets/img/wfhdog.gif'
-import { Button, Col, Form, Input, Row, Typography, Comment, Tooltip, Avatar, Modal, Card } from 'antd'
+import { Button, Col, Form, Row, Comment, Tooltip, Avatar, Modal, Card } from 'antd'
 import FinishModalContent from '../components/FinishModalContent'
+import TaskForm from '../components/TaskForm'
 import moment from 'moment'
 import asyncForEach from '../plugins/asyncForEach'
 import getUnixTime from '../plugins/getUnixTime'
@@ -10,19 +11,8 @@ import 'firebase/firestore'
 
 const db = firebase.firestore()
 
-const { Title } = Typography
-const layout = {
-  labelCol: { span: 8 },
-  wrapperCol: { span: 16 },
-}
-
 const Home = () => {
   const [form] = Form.useForm()
-  const [modalVisible, setModalVisible] = useState(false)
-  const [confirmLoading, setConfirmLoading] = useState(false)
-  const [postId, setPostId] = useState('')
-  const toggleConfirmLoading = (flg: boolean) => setConfirmLoading(flg)
-  const [declarations, setDeclarations] = useState<firebase.firestore.DocumentData[]>([])
   const onFinish = async (values: any) => {
     const newDeclarations = {
       text: values.declaration,
@@ -35,10 +25,14 @@ const Home = () => {
     declarationsCopy.unshift(newDeclarations)
     setDeclarations(declarationsCopy)
   }
+  const [modalVisible, setModalVisible] = useState(false)
+  const [confirmLoading, setConfirmLoading] = useState(false)
+  const [postId, setPostId] = useState('')
+  const toggleConfirmLoading = (flg: boolean) => setConfirmLoading(flg)
+  const [declarations, setDeclarations] = useState<firebase.firestore.DocumentData[]>([])
   useEffect(() => {
     const getDeclarations = async () => {
       const declarationsShot = await db.collection('declarations').orderBy('created', 'desc').get()
-      // const declarationsData = declarationsShot.docs.map(doc => ({ id: doc.id, ...doc.data() }))
       const declarationsData: any = []
       await asyncForEach(declarationsShot.docs, async doc => {
         const dec = doc.data()
@@ -64,19 +58,7 @@ const Home = () => {
       <Row>
         <Col span={6} offset={6}>
           <img src={wfhdog} alt="dog" />
-          <Typography>
-            <Title level={3}>What are going to do today?</Title>
-          </Typography>
-          <Form {...layout} form={form} name="nest-messages" onFinish={onFinish}>
-            <Form.Item name="declaration">
-              <Input.TextArea />
-            </Form.Item>
-            <Form.Item>
-              <Button type="primary" htmlType="submit">
-                Submit
-              </Button>
-            </Form.Item>
-          </Form>
+          <TaskForm form={form} onFinish={onFinish} />
           {declarations.map(declaration => (
             <>
               <Comment
